+++ samples/webcomponents/index.js
@@ -1,11 +1,13 @@
-import { installTwicPics, TwicImg, TwicVideo } from "@twicpics/components/webcomponents";
+import { installTwicPics, TwicImg, TwicVideo, TwicView } from "@twicpics/components/webcomponents";
 
 installTwicPics( {
     "domain": `https://demo.twic.pics`,
     "anticipation": 0.5,
     "step": 100,
     "env": `production`,
+    "class": `pipo`,
 } );
 
 customElements.define( `twic-img`, TwicImg );
 customElements.define( `twic-video`, TwicVideo );
+customElements.define( `twic-view`, TwicView );