+++ src/angular/base.component.ts
@@ -38,7 +38,7 @@ import {
     parseFocus,
     parseIntrinsic,
     parseMode,
-    parseNoLazyLoading,
+    parseEager,
     parsePlaceholder,
     parsePosition,
     parsePreTransform,
@@ -66,7 +66,7 @@ export class TwicBaseComponent implements OnInit, OnDestroy, OnChanges {
     @Input() focus: string = undefined;
     @Input() intrinsic: string = undefined;
     @Input() mode: Mode = undefined;
-    @Input() nolazyloading: boolean | string;
+    @Input() eager: boolean | string;
     @Input() placeholder: Placeholder = undefined;
     @Input() position: string = undefined;
     @Input() preTransform: string = undefined;
@@ -81,10 +81,10 @@ export class TwicBaseComponent implements OnInit, OnDestroy, OnChanges {
     _alt: string = undefined;
     _anchor: AnchorObject = undefined;
     _bot: string = undefined;
+    _eager: boolean;
     _focus: string = undefined;
     _intrinsic: string = undefined;
     _mode: Mode = undefined;
-    _nolazyloading: boolean;
     _placeholder: Placeholder = undefined;
     _position: string = undefined;
     _preTransform: string = undefined;
@@ -125,7 +125,7 @@ export class TwicBaseComponent implements OnInit, OnDestroy, OnChanges {
         this._focus = parseFocus( this.focus );
         this._intrinsic = parseIntrinsic( this.intrinsic );
         this._mode = parseMode( this.mode );
-        this._nolazyloading = parseNoLazyLoading( this.nolazyloading );
+        this._eager = parseEager( this.eager );
         this._placeholder = parsePlaceholder( this.placeholder, this.src );
         this._position = parsePosition( this.position );
         this._preTransform = parsePreTransform( this.preTransform );
@@ -143,10 +143,10 @@ export class TwicBaseComponent implements OnInit, OnDestroy, OnChanges {
             ...computeData(
                 this._anchor,
                 this._bot,
+                this._eager,
                 this._focus,
                 this._intrinsic,
                 this._mode,
-                this._nolazyloading,
                 this._preTransform,
                 this._src,
                 this._step